refactor: extract 3 inner types from TweedleUnlinkedParser (558→220 lines)
Extract ExpressionVisitor, StatementVisitor, and BinaryConstructor from
TweedleUnlinkedParser into package-private top-level classes. Promote 6
utility methods from private to package-private for access by extracted
visitors. Constructor-injected parser reference pattern preserves all
behavior with zero public API changes.

Resolves #714
