+++ core/tweedle/src/main/java/org/alice/tweedle/unlinked/BinaryConstructor.java
@@ -1,0 +1,9 @@
+package org.alice.tweedle.unlinked;
+
+import org.alice.tweedle.ast.BinaryExpression;
+import org.alice.tweedle.ast.TweedleExpression;
+
+@FunctionalInterface
+interface BinaryConstructor {
+  BinaryExpression newBinExp(TweedleExpression lhs, TweedleExpression rhs);
+}
