+++ core/tweedle/src/main/java/org/alice/tweedle/unlinked/TweedleUnlinkedParser.java
@@ -3,7 +3,6 @@ package org.alice.tweedle.unlinked;
 import org.alice.tweedle.*;
 import org.alice.tweedle.ast.*;
 import org.antlr.v4.runtime.*;
-import org.antlr.v4.runtime.tree.TerminalNode;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -20,11 +19,11 @@ public class TweedleUnlinkedParser {
   }
 
   TweedleStatement parseStatement(String sourceForExpression) {
-    return new StatementVisitor().visit(tweedleParserForSource(sourceForExpression).blockStatement());
+    return new StatementVisitor(this).visit(tweedleParserForSource(sourceForExpression).blockStatement());
   }
 
   TweedleExpression parseExpression(String sourceForExpression) {
-    return new ExpressionVisitor().visit(tweedleParserForSource(sourceForExpression).expression());
+    return new ExpressionVisitor(this).visit(tweedleParserForSource(sourceForExpression).expression());
   }
 
   private TweedleParser tweedleParserForSource(String source) {
@@ -123,7 +122,7 @@ public class TweedleUnlinkedParser {
       final String name = context.variableDeclarator().variableDeclaratorId().IDENTIFIER().getText();
       TweedleField property;
       if (context.variableDeclarator().variableInitializer() != null) {
-        ExpressionVisitor initVisitor = new ExpressionVisitor(type, true);
+        ExpressionVisitor initVisitor = new ExpressionVisitor(TweedleUnlinkedParser.this, type, true);
         TweedleExpression init = context.variableDeclarator().variableInitializer().accept(initVisitor);
         property = new TweedleField(modifiers, type, name, init);
       } else {
@@ -161,24 +160,24 @@ public class TweedleUnlinkedParser {
       return context.formalParameterList().optionalParameter().stream().map(field -> {
         final TweedleType type = getType(field.typeType());
         final String varName = field.variableDeclaratorId().IDENTIFIER().getText();
-        return new TweedleOptionalParameter(type, varName, field.accept(new ExpressionVisitor(type)));
+        return new TweedleOptionalParameter(type, varName, field.accept(new ExpressionVisitor(TweedleUnlinkedParser.this, type)));
       }).collect(toList());
     }
   }
 
-  private List<TweedleStatement> collectBlockStatements(List<TweedleParser.BlockStatementContext> contexts) {
-    StatementVisitor statementVisitor = new StatementVisitor();
+  List<TweedleStatement> collectBlockStatements(List<TweedleParser.BlockStatementContext> contexts) {
+    StatementVisitor statementVisitor = new StatementVisitor(this);
     return contexts.stream().map(stmt -> stmt.accept(statementVisitor)).collect(toList());
   }
 
-  private Map<String, TweedleExpression> visitLabeledArguments(TweedleParser.LabeledExpressionListContext context) {
+  Map<String, TweedleExpression> visitLabeledArguments(TweedleParser.LabeledExpressionListContext context) {
     if (context == null) {
       return Collections.emptyMap();
     }
 
     List<TweedleParser.LabeledExpressionContext> argumentContexts = context.labeledExpression();
     Map<String, TweedleExpression> arguments = new HashMap<>(argumentContexts.size());
-    final ExpressionVisitor visitor = new ExpressionVisitor();
+    final ExpressionVisitor visitor = new ExpressionVisitor(this);
     for (TweedleParser.LabeledExpressionContext arg : argumentContexts) {
       TweedleExpression argValue = arg.expression().accept(visitor);
       arguments.put(arg.IDENTIFIER().getText(), argValue);
@@ -186,18 +185,18 @@ public class TweedleUnlinkedParser {
     return arguments;
   }
 
-  private List<TweedleExpression> visitUnlabeledArguments(TweedleParser.UnlabeledExpressionListContext listContext, ExpressionVisitor expressionVisitor) {
+  List<TweedleExpression> visitUnlabeledArguments(TweedleParser.UnlabeledExpressionListContext listContext, ExpressionVisitor expressionVisitor) {
     return listContext == null ? new ArrayList<>() : listContext.expression().stream().map(a -> a.accept(expressionVisitor)).collect(toList());
   }
 
-  private TweedleType getTypeOrVoid(TweedleParser.TypeTypeOrVoidContext context) {
+  TweedleType getTypeOrVoid(TweedleParser.TypeTypeOrVoidContext context) {
     if (context.VOID() != null) {
       return TweedleVoidType.VOID;
     }
     return getType(context.typeType());
   }
 
-  private TweedleType getType(TweedleParser.TypeTypeContext context) {
+  TweedleType getType(TweedleParser.TypeTypeContext context) {
     TweedleType baseType = context.classType() != null ? getTypeReference(context.classType().getText()) : getPrimitiveType(context.primitiveType().getText());
     if (context.getChildCount() > 1 && baseType != null) {
       return new TweedleArrayType(baseType);
@@ -205,11 +204,11 @@ public class TweedleUnlinkedParser {
     return baseType;
   }
 
-  private TweedleTypeReference getTypeReference(String typeName) {
+  TweedleTypeReference getTypeReference(String typeName) {
     return new TweedleTypeReference(typeName);
   }
 
-  private TweedlePrimitiveType getPrimitiveType(String typeName) {
+  TweedlePrimitiveType getPrimitiveType(String typeName) {
     for (TweedlePrimitiveType prim : TweedleTypes.PRIMITIVE_TYPES) {
       if (prim.getName().equals(typeName)) {
         return prim;
@@ -218,341 +217,4 @@ public class TweedleUnlinkedParser {
     return null;
   }
 
-  private class ExpressionVisitor extends TweedleParserBaseVisitor<TweedleExpression> {
-    private TweedleType expectedType;
-    private final boolean allowPrimitiveNull;
-
-    ExpressionVisitor() {
-      this(null);
-    }
-
-    ExpressionVisitor(TweedleType expectedType) {
-      this(expectedType, false);
-    }
-
-    ExpressionVisitor(TweedleType expectedType, boolean allowPrimitiveNull) {
-      this.expectedType = expectedType;
-      this.allowPrimitiveNull = allowPrimitiveNull;
-    }
-
-    @Override
-    public TweedleExpression visitPrimary(TweedleParser.PrimaryContext context) {
-      if (context.expression() != null) {
-        // Parenthesized child expression
-        return this.visitExpression(context.expression());
-      }
-      if (context.THIS() != null) {
-        return new ThisExpression();
-      }
-      if (context.IDENTIFIER() != null) {
-        return new IdentifierReference(context.IDENTIFIER().getText());
-      }
-      // TODO parse super superSuffix
-
-      // Visit children to handle literals
-      return super.visitPrimary(context);
-    }
-
-    @Override
-    public TweedleExpression visitLiteral(TweedleParser.LiteralContext context) {
-      TerminalNode wholeNumber = context.DECIMAL_LITERAL();
-      if (wholeNumber != null) {
-        int value = Integer.parseInt(wholeNumber.getSymbol().getText());
-        return TweedleTypes.WHOLE_NUMBER.createValue(value);
-      }
-
-      TerminalNode flt = context.FLOAT_LITERAL();
-      if (flt != null) {
-        double value = Double.parseDouble(flt.getSymbol().getText());
-        return TweedleTypes.DECIMAL_NUMBER.createValue(value);
-      }
-
-      if (context.NULL_LITERAL() != null) {
-        return TweedleNull.NULL;
-      }
-
-      TerminalNode bool = context.BOOL_LITERAL();
-      if (bool != null) {
-        boolean value = Boolean.parseBoolean(bool.getSymbol().getText());
-        return TweedleTypes.BOOLEAN.createValue(value);
-      }
-
-      TerminalNode str = context.STRING_LITERAL();
-      if (str != null) {
-        final String quotedString = str.getSymbol().getText();
-        return TweedleTypes.TEXT_STRING.createValue(quotedString.substring(1, quotedString.length() - 1));
-      }
-
-      return super.visitLiteral(context);
-    }
-
-    @Override
-    public TweedleExpression visitExpression(TweedleParser.ExpressionContext context) {
-      TweedleExpression expression = buildExpression(context);
-      if (!isExpectedTypeCompatible(expression)) {
-        throw new RuntimeException("Had been expecting expression of type " + expectedType + ", but it is typed as " + expression.getType());
-      }
-      return expression;
-    }
-
-    private boolean isExpectedTypeCompatible(TweedleExpression expression) {
-      if (expectedType == null || expression == null || expression.getType() == null) {
-        return true;
-      }
-      if (expression instanceof TweedleNull) {
-        return expectedType == TweedleTypes.TEXT_STRING
-            || !(expectedType instanceof TweedlePrimitiveType<?>)
-            || allowPrimitiveNull;
-      }
-      return expectedType.willAcceptValueOfType(expression.getType()) || expression.getType().willAcceptValueOfType(expectedType);
-    }
-
-    private TweedleExpression buildExpression(TweedleParser.ExpressionContext context) {
-      Token prefix = context.prefix;
-      Token operation = context.bop;
-
-      if (prefix != null) {
-        switch (prefix.getText()) {
-        case "+":
-          // A positive number, or at least not changing the sign. Send along child.
-          return getFirstExpression(TweedleTypes.NUMBER, context);
-        case "-":
-          // A negative number, or a sign flip. Send along negated child.
-          return getNegativeOfExpression(getFirstExpression(TweedleTypes.NUMBER, context));
-        case "!":
-          return new LogicalNotExpression(getFirstExpression(TweedleTypes.BOOLEAN, context));
-        default:
-          throw new RuntimeException("Unrecognized prefix operation: " + prefix.getText());
-        }
-      } else if (operation != null) {
-        switch (operation.getText()) {
-        case ".":
-          return fieldOrMethodRef(context);
-        case "==":
-          return binaryExpression(EqualToExpression::new, null, context); //XxY=>B
-        case "!=":
-          return binaryExpression(NotEqualToExpression::new, null, context); //XxY=>B
-        case "..":
-          return binaryExpression(StringConcatenationExpression::new, null, context); //XxY=>B
-        case "*":
-          return binaryExpression(MultiplicationExpression::new, TweedleTypes.NUMBER, context);
-        case "/":
-          return binaryExpression(DivisionExpression::new, TweedleTypes.NUMBER, context);
-        case "%":
-          return binaryExpression(ModuloExpression::new, TweedleTypes.WHOLE_NUMBER, context);
-        case "+":
-          return binaryExpression(AdditionExpression::new, TweedleTypes.NUMBER, context);
-        case "-":
-          return binaryExpression(SubtractionExpression::new, TweedleTypes.NUMBER, context);
-        case "<=":
-          return binaryExpression(LessThanOrEqualExpression::new, TweedleTypes.NUMBER, context);
-        case ">=":
-          return binaryExpression(GreaterThanOrEqualExpression::new, TweedleTypes.NUMBER, context);
-        case ">":
-          return binaryExpression(GreaterThanExpression::new, TweedleTypes.NUMBER, context);
-        case "<":
-          return binaryExpression(LessThanExpression::new, TweedleTypes.NUMBER, context);
-        case "&&":
-          return binaryExpression(LogicalAndExpression::new, TweedleTypes.BOOLEAN, context);
-        case "||":
-          return binaryExpression(LogicalOrExpression::new, TweedleTypes.BOOLEAN, context);
-        case "<-":
-          List<TweedleExpression> expressions = getTypedExpressions(context, null);
-          return new AssignmentExpression(expressions.getFirst(), expressions.get(1));
-        default:
-          throw new RuntimeException("No such operation as " + operation.getText());
-        }
-      } else if (context.bracket != null) {
-        TweedleExpression array = context.expression(0).accept(new ExpressionVisitor(new TweedleArrayType()));
-        TweedleExpression index = context.expression(1).accept(new ExpressionVisitor(TweedleTypes.WHOLE_NUMBER));
-        return new ArrayIndexExpression(((TweedleArrayType) array.getType()).getValueType(), array, index);
-      } else if (context.lambdaCall() != null) {
-        TweedleExpression lambdaSourceExp = getFirstExpression(null, context);
-        if (context.lambdaCall().unlabeledExpressionList() == null) {
-          return new LambdaEvaluation(lambdaSourceExp);
-        } else {
-          final List<TweedleExpression> elements = visitUnlabeledArguments(context.lambdaCall().unlabeledExpressionList(), new ExpressionVisitor());
-          return new LambdaEvaluation(lambdaSourceExp, elements);
-        }
-      }
-
-      // This will handle primary & lambda
-      return visitChildren(context);
-    }
-
-    @Override
-    public TweedleExpression visitLambdaExpression(TweedleParser.LambdaExpressionContext ctx) {
-      List<TweedleRequiredParameter> parameters = lambdaParameters(ctx.lambdaParameters());
-      List<TweedleStatement> stmts = collectBlockStatements(ctx.block().blockStatement());
-      return new LambdaExpression(parameters, stmts);
-    }
-
-    private List<TweedleRequiredParameter> lambdaParameters(TweedleParser.LambdaParametersContext context) {
-      return context.requiredParameter().stream().map(field -> new TweedleRequiredParameter(getType(field.typeType()), field.variableDeclaratorId().IDENTIFIER().getText())).collect(toList());
-    }
-
-    @Override
-    public TweedleExpression visitMethodCall(TweedleParser.MethodCallContext ctx) {
-
-      return new MethodCallExpression(new ThisExpression(), ctx.IDENTIFIER().getText(), visitLabeledArguments(ctx.labeledExpressionList()), false);
-    }
-
-    public @Override
-    TweedleExpression visitSuperSuffix(TweedleParser.SuperSuffixContext context) {
-
-      if (context.IDENTIFIER() != null) {
-        if (context.arguments() != null) {
-          return new MethodCallExpression(new SuperExpression(), context.IDENTIFIER().getText(), visitLabeledArguments(context.arguments().labeledExpressionList()));
-        } else {
-          return new FieldAccess(new SuperExpression(), context.IDENTIFIER().getText());
-        }
-      } else if (context.arguments() != null) {
-        return new Instantiation(new SuperExpression(), visitLabeledArguments(context.arguments().labeledExpressionList()));
-      }
-      throw new RuntimeException("Super suffix could not be constructed.");
-    }
-
-    public @Override
-    TweedleExpression visitCreator(TweedleParser.CreatorContext context) {
-      String typeName = context.createdName().getText();
-      final TweedleParser.ArrayCreatorRestContext arrayDetails = context.arrayCreatorRest();
-      if (arrayDetails != null) {
-        TweedlePrimitiveType prim = getPrimitiveType(typeName);
-        TweedleType memberType = prim == null ? getTypeReference(typeName) : prim;
-        TweedleArrayType arrayType = new TweedleArrayType(memberType);
-        if (arrayDetails.arrayInitializer() != null) {
-          final List<TweedleExpression> elements = visitUnlabeledArguments(arrayDetails.arrayInitializer().unlabeledExpressionList(), new ExpressionVisitor(memberType));
-          return new TweedleArrayInitializer(arrayType, elements);
-        } else {
-          return new TweedleArrayInitializer(arrayType, arrayDetails.expression().accept(new ExpressionVisitor()));
-        }
-      } else {
-        TweedleTypeReference typeRef = getTypeReference(typeName);
-        TweedleParser.LabeledExpressionListContext argsContext = context.classCreatorRest().arguments().labeledExpressionList();
-        Map<String, TweedleExpression> arguments = visitLabeledArguments(argsContext);
-        return new Instantiation(typeRef, arguments);
-      }
-    }
-
-    private TweedleExpression fieldOrMethodRef(TweedleParser.ExpressionContext context) {
-      // Use untyped expression visitor for target
-      TweedleExpression target = context.expression(0).accept(new ExpressionVisitor());
-      if (context.IDENTIFIER() != null) {
-        return new FieldAccess(target, context.IDENTIFIER().getText());
-      }
-      if (context.methodCall() != null) {
-        return new MethodCallExpression(
-            target,
-            context.methodCall().IDENTIFIER().getText(),
-            visitLabeledArguments(context.methodCall().labeledExpressionList()));
-      }
-      throw new RuntimeException("Unexpected details on context " + context);
-    }
-
-    private TweedleExpression getNegativeOfExpression(TweedleExpression exp) {
-      final NegativeExpression negativeExpression = new NegativeExpression(exp);
-      if (exp instanceof TweedlePrimitiveValue) {
-        return negativeExpression.evaluate(null);
-      }
-      return negativeExpression;
-    }
-
-    private TweedleExpression getFirstExpression(TweedlePrimitiveType type, TweedleParser.ExpressionContext context) {
-      return getTypedExpressions(context, type).getFirst();
-    }
-
-    private TweedleExpression binaryExpression(BinaryConstructor constructor, TweedlePrimitiveType type, TweedleParser.ExpressionContext context) {
-      List<TweedleExpression> expressions = getTypedExpressions(context, type);
-      return constructor.newBinExp(expressions.getFirst(), expressions.get(1));
-    }
-
-    private List<TweedleExpression> getTypedExpressions(TweedleParser.ExpressionContext context, TweedleType type) {
-      final ExpressionVisitor visitor = new ExpressionVisitor(type);
-      return context.expression().stream().map(exp -> exp.accept(visitor)).collect(toList());
-    }
-  }
-
-  public interface BinaryConstructor {
-    BinaryExpression newBinExp(TweedleExpression lhs, TweedleExpression rhs);
-  }
-
-  private class StatementVisitor extends TweedleParserBaseVisitor<TweedleStatement> {
-
-    @Override
-    public TweedleStatement visitLocalVariableDeclaration(TweedleParser.LocalVariableDeclarationContext context) {
-      TweedleType type = getType(context.typeType());
-      final String name = context.variableDeclarator().variableDeclaratorId().IDENTIFIER().getText();
-      TweedleLocalVariable decl;
-      if (context.variableDeclarator().variableInitializer() != null) {
-        ExpressionVisitor initVisitor = new ExpressionVisitor(type);
-        TweedleExpression init = context.variableDeclarator().variableInitializer().accept(initVisitor);
-        decl = new TweedleLocalVariable(type, name, init);
-      } else {
-        decl = new TweedleLocalVariable(type, name);
-      }
-      return new LocalVariableDeclaration(context.CONSTANT() != null, decl);
-    }
-
-    @Override
-    public TweedleStatement visitBlockStatement(TweedleParser.BlockStatementContext context) {
-      if (context.NODE_DISABLE() != null) {
-        TweedleStatement stmt = context.blockStatement().accept(this);
-        stmt.disable();
-        return stmt;
-      }
-      if (context.localVariableDeclaration() != null) {
-        return context.localVariableDeclaration().accept(this);
-      }
-      return super.visitBlockStatement(context);
-    }
-
-    @Override
-    public TweedleStatement visitStatement(TweedleParser.StatementContext context) {
-      if (context.COUNT_UP_TO() != null) {
-        return new CountUpLoop(context.IDENTIFIER().getText(), context.expression().accept(new ExpressionVisitor(TweedleTypes.WHOLE_NUMBER)), collectBlockStatements(context.block(0).blockStatement()));
-      }
-      if (context.IF() != null) {
-        TweedleExpression condition = context.parExpression().expression().accept(new ExpressionVisitor(TweedleTypes.BOOLEAN));
-        List<TweedleStatement> thenBlock = collectBlockStatements(context.block(0).blockStatement());
-        List<TweedleStatement> elseBlock = context.ELSE() != null ? collectBlockStatements(context.block(1).blockStatement()) : new ArrayList<>();
-        return new ConditionalStatement(condition, thenBlock, elseBlock);
-      }
-      if (context.forControl() != null) {
-        final TweedleType valueType = getType(context.forControl().typeType());
-        final TweedleArrayType arrayType = new TweedleArrayType(valueType);
-        final TweedleLocalVariable loopVar = new TweedleLocalVariable(valueType, context.forControl().variableDeclaratorId().getText());
-        final TweedleExpression loopValues = context.forControl().expression().accept(new ExpressionVisitor(arrayType));
-        final List<TweedleStatement> statements = collectBlockStatements(context.block(0).blockStatement());
-        if (context.FOR_EACH() != null) {
-          return new ForEachLoop(loopVar, loopValues, statements);
-        }
-        if (context.EACH_TOGETHER() != null) {
-          return new ForEachTogether(loopVar, loopValues, statements);
-        }
-        throw new RuntimeException("Found a forControl in a statement where it was not expected: " + context);
-      }
-      if (context.WHILE() != null) {
-        return new WhileLoop(context.parExpression().expression().accept(new ExpressionVisitor(TweedleTypes.BOOLEAN)), collectBlockStatements(context.block(0).blockStatement()));
-      }
-      if (context.DO_IN_ORDER() != null) {
-        return new DoInOrder(collectBlockStatements(context.block(0).blockStatement()));
-      }
-      if (context.DO_TOGETHER() != null) {
-        return new DoTogether(collectBlockStatements(context.block(0).blockStatement()));
-      }
-      if (context.RETURN() != null) {
-        if (context.expression() != null) {
-          return new ReturnStatement(context.expression().accept(new ExpressionVisitor()));
-        } else {
-          return new ReturnStatement();
-        }
-      }
-      TweedleParser.ExpressionContext expContext = context.statementExpression;
-      if (expContext != null) {
-        return new ExpressionStatement(context.expression().accept(new ExpressionVisitor()));
-      }
-      throw new RuntimeException("Found a statement that was not expected: " + context);
-    }
-
-  }
 }
